Reworked getDefaults and getValidation to use attributes
The "required" attribute isn't available in model.definitions, but it is present in model._attributes.  Switching to this data results in this information getting stored in the BackboneModels.

Also, defaults were not returning as it seems the code was written for data in a different structure than either definitions or _attributes has it stored.  Here is an example of how I found each of the getDefaults methods to work: http://jsfiddle.net/hsj73onw/  I switched to _attributes for consistency sake though it didn't seem to be required as the "defaultsTo" property appears in both places.